backend/pyproject.toml
[project]
name = "fast-feast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt<4.0.0",
    "cloudinary>=1.44.2",
    "fastapi[standard]>=0.136.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.38",
    "stripe>=15.0.1",
]
